ext_symlink.sh: Remove channels if it exists This script was trying to remove AT channels (/dev/CHNPTY*) even if it did not exist. And IDK how strict is init when it comes to run the services, in this case, it issued a SIGKILL, causing the script to not execute properly, thereby not creating proper symlinks for AT channels. Which is why RIL did not work.
This patch fixes the issue, at expense of performance, RIL does take a while to load.

+++ proprietary/bin/ext_symlink.sh
@@ -4,11 +4,15 @@ term="/dev/pts/* "
 
 if [ "$1" = "-p" ]; then
 	link=`getprop sys.symlink.pty`
-	rm ${link##${term}}
+	if [ -f ${link##${term}} ]; then
+		rm ${link##${term}}
+	fi
 	ln -s $link;
 	setprop sys.symlink.notify 0
 elif [ "$1" = "-u" ]; then
 	link=`getprop sys.symlink.umts_router`
-	rm ${link##${term}}
+	if [ -f ${link##${term}} ]; then
+		rm ${link##${term}}
+	fi
 	ln -s $link;
 fi
